move tests to own directory
