+++ tests/crit_app/util/api/test_data/xiv_gear_test_data_2.json
