+++ tests/fflogs_rotation/dawntrail/healer/test_ast.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+from ffxiv_stats.jobs import Healer
 from pandas.testing import assert_frame_equal
 
 from crit_app.job_data.encounter_data import encounter_phases
@@ -13,44 +14,43 @@ from fflogs_rotation.job_data.data import (
 )
 from fflogs_rotation.rotation import RotationTable
 from fflogs_rotation.tests.config import headers
-from ffxiv_stats.jobs import Healer
 
 
-class TestWhiteMageActions:
-    """Tests for White Mage job actions and rotations.
+class TestAstrologianActions:
+    """Tests for Astrologian job actions and rotations.
 
-    Based off m2s
-    https://www.fflogs.com/reports/8k7xvmRyg6AVMTW1?fight=12
+    Based of Reality's speed kill for m1s
+    https://www.fflogs.com/reports/B3gxKdn4j1W8NML9#fight=3
     """
 
     def setup_method(self):
         """Setup method to initialize common variables and RotationTable."""
-        self.report_id = "8k7xvmRyg6AVMTW1"
-        self.fight_id = 12
+        self.report_id = "B3gxKdn4j1W8NML9"
+        self.fight_id = 3
         self.level = 100
         self.phase = 0
-        self.player_id = 24
-        self.pet_ids = None
+        self.player_id = 8
+        self.pet_ids = [12]
 
-        self.main_stat = 5052
+        self.main_stat = 5091
         self.pet_attack_power = self.main_stat // 1.05
-        self.det = 2803
-        self.speed = 420
+        self.det = 3043
+        self.speed = 528
 
-        self.crt = 3147
-        self.dh = 1320
+        self.crt = 2922
+        self.dh = 1050
         self.wd = 146
 
-        self.delay = 2.8
+        self.delay = 3.2
         self.medication = 392
 
-        self.t = 511
+        self.t = 393.947
 
-        self.whm_rotation_7_05 = RotationTable(
+        self.ast_rotation_7_05 = RotationTable(
             headers,
             self.report_id,
             self.fight_id,
-            "WhiteMage",
+            "Astrologian",
             self.player_id,
             self.crt,
             self.dh,
@@ -68,7 +68,7 @@ class TestWhiteMageActions:
             pet_ids=self.pet_ids,
         )
 
-        self.whm_analysis_7_05 = Healer(
+        self.ast_analysis_7_05 = Healer(
             self.main_stat,
             400,
             self.det,
@@ -83,30 +83,33 @@ class TestWhiteMageActions:
         # self.black_cat_ast.attach_rotation(self.black_cat_rotation.rotation_df, self.t)
 
     @pytest.fixture
-    def expected_whm_7_05_action_counts(self) -> pd.DataFrame:
+    def expected_ast_7_05_action_counts(self) -> pd.DataFrame:
         """Fixture providing expected action count data."""
         return (
             pd.DataFrame(
-                [
-                    {"base_action": "Dia (tick)", "n": 168},
-                    {"base_action": "Glare III", "n": 147},
-                    {"base_action": "Dia", "n": 18},
-                    {"base_action": "Glare IV", "n": 15},
-                    {"base_action": "Assize", "n": 13},
-                    {"base_action": "Afflatus Misery", "n": 8},
-                ]
+                {
+                    "base_action": [
+                        "Fall Malefic",
+                        "Combust III (tick)",
+                        "Oracle",
+                        "Stellar Explosion (Pet)",
+                        "Lord of Crowns",
+                        "Macrocosmos",
+                    ],
+                    "n": [138, 127, 4, 7, 4, 2],
+                }
             )
             .sort_values(["n", "base_action"], ascending=[False, True])
             .reset_index(drop=True)
         )
 
-    def test_whm_7_05_action_counts(
-        self, expected_whm_7_05_action_counts: pd.DataFrame
+    def test_ast_7_05_action_counts(
+        self, expected_ast_7_05_action_counts: pd.DataFrame
     ):
-        """Test that action counts match expected values against FFLogs."""
+        """Test that action counts match expected values for Black Cat log."""
         # Arrange
         actual_counts = (
-            self.whm_rotation_7_05.rotation_df.groupby("base_action")
+            self.ast_rotation_7_05.rotation_df.groupby("base_action")
             .sum("n")
             .reset_index()[["base_action", "n"]]
             .sort_values(["n", "base_action"], ascending=[False, True])
@@ -114,24 +117,24 @@ class TestWhiteMageActions:
         )
 
         # Assert
-        assert_frame_equal(actual_counts, expected_whm_7_05_action_counts)
+        assert_frame_equal(actual_counts, expected_ast_7_05_action_counts)
 
 
 class TestAstrologianMultiTargetActions:
     """Tests for Astrologian job actions and rotations with an emphasis on mutli-target.
 
-    Based off https://www.fflogs.com/reports/v2ZnTdW67wfXN3gQ?fight=11&type=damage-done&source=5
+    Based off https://www.fflogs.com/reports/ZfnF8AqRaBbzxW3w?fight=5
     """
 
     def setup_method(self):
         """Setup method to initialize common variables and RotationTable."""
-        self.report_id = "v2ZnTdW67wfXN3gQ"
-        self.fight_id = 11
+        self.report_id = "ZfnF8AqRaBbzxW3w"
+        self.fight_id = 5
         self.level = 100
         self.phase = 2
-        self.player_id = 5
-        self.pet_ids = None
-        self.excluded_enemy_ids = [35]
+        self.player_id = 27
+        self.pet_ids = [30]
+        self.excluded_enemy_ids = [52]
 
         self.main_stat = 5129
         self.pet_attack_power = self.main_stat // 1.05
@@ -147,11 +150,11 @@ class TestAstrologianMultiTargetActions:
 
         self.t = 1
 
-        self.whm_rotation_7_1_fru_p2 = RotationTable(
+        self.ast_rotation_7_1_fru_p2 = RotationTable(
             headers,
             self.report_id,
             self.fight_id,
-            "WhiteMage",
+            "Astrologian",
             self.player_id,
             self.crt,
             self.dh,
@@ -170,7 +173,7 @@ class TestAstrologianMultiTargetActions:
             excluded_enemy_ids=self.excluded_enemy_ids,
         )
 
-        self.whm_analysis_7_1_fru_p2 = Healer(
+        self.ast_analysis_7_1_fru_p2 = Healer(
             self.main_stat,
             400,
             self.det,
@@ -185,11 +188,11 @@ class TestAstrologianMultiTargetActions:
         # self.black_cat_ast.attach_rotation(self.black_cat_rotation.rotation_df, self.t)
 
         ####### P4
-        self.whm_rotation_7_1_fru_p4 = RotationTable(
+        self.ast_rotation_7_1_fru_p4 = RotationTable(
             headers,
             self.report_id,
             self.fight_id,
-            "WhiteMage",
+            "Astrologian",
             self.player_id,
             self.crt,
             self.dh,
@@ -209,30 +212,32 @@ class TestAstrologianMultiTargetActions:
         )
 
     @pytest.fixture
-    def expected_whm_fru_p2_7_1_action_counts(self) -> pd.DataFrame:
+    def expected_ast_fru_p2_7_1_action_counts(self) -> pd.DataFrame:
         """Fixture providing expected action count data."""
         return (
             pd.DataFrame(
-                [
-                    {"base_action": "Dia (tick)", "n": 44},
-                    {"base_action": "Glare III", "n": 40},
-                    {"base_action": "Assize", "n": 6},
-                    {"base_action": "Glare IV", "n": 6},
-                    {"base_action": "Dia", "n": 5},
-                    {"base_action": "Afflatus Misery", "n": 3},
-                ]
+                {
+                    "base_action": [
+                        "Fall Malefic",
+                        "Combust III (tick)",
+                        "Oracle",
+                        "Lord of Crowns",
+                        "Stellar Explosion (Pet)",
+                    ],
+                    "n": [49, 49, 2, 3, 4],
+                }
             )
             .sort_values(["n", "base_action"], ascending=[False, True])
             .reset_index(drop=True)
         )
 
-    def test_whm_fru_p2_action_totals(
-        self, expected_whm_fru_p2_7_1_action_counts: pd.DataFrame
+    def test_ast_fru_p2_action_totals(
+        self, expected_ast_fru_p2_7_1_action_counts: pd.DataFrame
     ):
         """Test that action counts match expected values."""
         # Arrange
         actual_counts = (
-            self.whm_rotation_7_1_fru_p2.rotation_df.groupby("base_action")
+            self.ast_rotation_7_1_fru_p2.rotation_df.groupby("base_action")
             .sum("n")
             .reset_index()[["base_action", "n"]]
             .sort_values(["n", "base_action"], ascending=[False, True])
@@ -240,33 +245,37 @@ class TestAstrologianMultiTargetActions:
         )
 
         # Assert
-        assert_frame_equal(actual_counts, expected_whm_fru_p2_7_1_action_counts)
+        assert_frame_equal(actual_counts, expected_ast_fru_p2_7_1_action_counts)
 
     @pytest.fixture
-    def expected_whm_fru_p4_7_1_action_counts(self) -> pd.DataFrame:
+    def expected_ast_fru_p4_7_1_action_counts(self) -> pd.DataFrame:
         """Fixture providing expected action count data."""
         return (
             pd.DataFrame(
-                [
-                    {"base_action": "Dia (tick)", "n": 45},
-                    {"base_action": "Glare III", "n": 24},
-                    {"base_action": "Assize", "n": 7},
-                    {"base_action": "Afflatus Misery", "n": 6},
-                    {"base_action": "Glare IV", "n": 6},
-                    {"base_action": "Dia", "n": 6},
-                ]
+                {
+                    "base_action": [
+                        "Combust III (tick)",
+                        "Gravity II",
+                        "Fall Malefic",
+                        "Stellar Explosion (Pet)",
+                        "Oracle",
+                        "Lord of Crowns",
+                        "Macrocosmos",
+                    ],
+                    "n": [48, 20, 14, 3, 2, 2, 2],
+                }
             )
             .sort_values(["n", "base_action"], ascending=[False, True])
             .reset_index(drop=True)
         )
 
-    def test_whm_fru_p4_action_totals(
-        self, expected_whm_fru_p4_7_1_action_counts: pd.DataFrame
+    def test_ast_fru_p4_action_totals(
+        self, expected_ast_fru_p4_7_1_action_counts: pd.DataFrame
     ):
         """Test that action counts match expected values."""
         # Arrange
         actual_counts = (
-            self.whm_rotation_7_1_fru_p4.rotation_df.groupby("base_action")
+            self.ast_rotation_7_1_fru_p4.rotation_df.groupby("base_action")
             .sum("n")
             .reset_index()[["base_action", "n"]]
             .sort_values(["n", "base_action"], ascending=[False, True])
@@ -274,4 +283,4 @@ class TestAstrologianMultiTargetActions:
         )
 
         # Assert
-        assert_frame_equal(actual_counts, expected_whm_fru_p4_7_1_action_counts)
+        assert_frame_equal(actual_counts, expected_ast_fru_p4_7_1_action_counts)
