+++ tests/fflogs_rotation/dawntrail/melee/test_sam.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+from ffxiv_stats.jobs import Melee
 from pandas.testing import assert_frame_equal
 
 from crit_app.job_data.encounter_data import encounter_phases
@@ -13,44 +14,42 @@ from fflogs_rotation.job_data.data import (
 )
 from fflogs_rotation.rotation import RotationTable
 from fflogs_rotation.tests.config import headers
-from ffxiv_stats.jobs import Melee
 
 
-class TestNinjaActions:
-    """Tests for Ninja job actions and rotations.
+class TestSamuraiActions:
+    """Tests for Samurai job actions and rotations.
 
-    Based of Reality's speed kill for m1s
-    https://www.fflogs.com/reports/B3gxKdn4j1W8NML9#fight=3
+    Based off https://www.fflogs.com/reports/Lw7PgrFVcA1GNqnh?fight=6
     """
 
     def setup_method(self):
         """Setup method to initialize common variables and RotationTable."""
-        self.report_id = "B3gxKdn4j1W8NML9"
-        self.fight_id = 3
+        self.report_id = "Lw7PgrFVcA1GNqnh"
+        self.fight_id = 6
         self.level = 100
         self.phase = 0
-        self.player_id = 7
-        self.pet_ids = [15]
+        self.player_id = 3
+        self.pet_ids = None
 
-        self.main_stat = 5105
+        self.main_stat = 5114
         self.pet_attack_power = self.main_stat // 1.05
-        self.det = 2387
+        self.det = 1931
         self.speed = 420
 
-        self.crt = 3173
-        self.dh = 1482
+        self.crt = 3103
+        self.dh = 2095
         self.wd = 146
 
-        self.delay = 2.56
+        self.delay = 2.64
         self.medication = 392
 
-        self.t = 393.947
+        self.t = 515
 
-        self.nin_rotation_7_05 = RotationTable(
+        self.sam_rotation_7_05 = RotationTable(
             headers,
             self.report_id,
             self.fight_id,
-            "Ninja",
+            "Samurai",
             self.player_id,
             self.crt,
             self.dh,
@@ -68,7 +67,7 @@ class TestNinjaActions:
             pet_ids=self.pet_ids,
         )
 
-        self.nin_analysis_7_05 = Melee(
+        self.sam_analysis_7_05 = Melee(
             self.main_stat,
             self.det,
             self.speed,
@@ -76,7 +75,7 @@ class TestNinjaActions:
             self.dh,
             self.wd,
             self.delay,
-            "Ninja",
+            "Samurai",
             self.pet_attack_power,
             level=self.level,
         )
@@ -88,41 +87,41 @@ class TestNinjaActions:
         return (
             pd.DataFrame(
                 [
-                    {"base_action": "Attack", "n": 167},
-                    {"base_action": "Spinning Edge", "n": 38},
-                    {"base_action": "Gust Slash", "n": 37},
-                    {"base_action": "Aeolian Edge", "n": 26},
-                    {"base_action": "Bhavacakra", "n": 22},
-                    {"base_action": "Dream Within a Dream", "n": 21},
-                    {"base_action": "Fleeting Raiju", "n": 18},
-                    {"base_action": "Raiton", "n": 18},
-                    {"base_action": "Armor Crush", "n": 13},
-                    {"base_action": "Fleeting Raiju (Pet)", "n": 11},
-                    {"base_action": "Hyosho Ranryu", "n": 7},
-                    {"base_action": "Kunai's Bane", "n": 7},
-                    {"base_action": "Suiton", "n": 11},
-                    {"base_action": "Tenri Jindo", "n": 4},
-                    {"base_action": "Zesho Meppo", "n": 4},
-                    {"base_action": "Fuma Shuriken", "n": 4},
-                    {"base_action": "Dokumori", "n": 4},
-                    {"base_action": "Phantom Kamaitachi (Pet)", "n": 3},
-                    {"base_action": "Aeolian Edge (Pet)", "n": 3},
-                    {"base_action": "Gust Slash (Pet)", "n": 3},
-                    {"base_action": "Spinning Edge (Pet)", "n": 2},
-                    {"base_action": "Armor Crush (Pet)", "n": 1},
+                    {"base_action": "Attack", "n": 202},
+                    {"base_action": "Higanbana (tick)", "n": 159},
+                    {"base_action": "Gyofu", "n": 53},
+                    {"base_action": "Hissatsu: Shinten", "n": 51},
+                    {"base_action": "Gekko", "n": 32},
+                    {"base_action": "Kasha", "n": 27},
+                    {"base_action": "Yukikaze", "n": 27},
+                    {"base_action": "Jinpu", "n": 16},
+                    {"base_action": "Kaeshi: Setsugekka", "n": 15},
+                    {"base_action": "Midare Setsugekka", "n": 15},
+                    {"base_action": "Shoha", "n": 13},
+                    {"base_action": "Shifu", "n": 12},
+                    {"base_action": "Tendo Kaeshi Setsugekka", "n": 11},
+                    {"base_action": "Tendo Setsugekka", "n": 11},
+                    {"base_action": "Hissatsu: Senei", "n": 9},
+                    {"base_action": "Higanbana", "n": 8},
+                    {"base_action": "Hissatsu: Gyoten", "n": 7},
+                    {"base_action": "Kaeshi: Namikiri", "n": 5},
+                    {"base_action": "Ogi Namikiri", "n": 5},
+                    {"base_action": "Zanshin", "n": 5},
+                    {"base_action": "Enpi", "n": 2},
+                    {"base_action": "Hissatsu: Yaten", "n": 1},
                 ]
             )
             .sort_values(["n", "base_action"], ascending=[False, True])
             .reset_index(drop=True)
         )
 
-    def test_nin_7_05_action_counts(
+    def test_sam_7_05_action_counts(
         self, expected_sam_7_05_action_counts: pd.DataFrame
     ):
         """Test that action counts match expected values for Black Cat log."""
         # Arrange
         actual_counts = (
-            self.nin_rotation_7_05.rotation_df.groupby("base_action")
+            self.sam_rotation_7_05.rotation_df.groupby("base_action")
             .sum("n")
             .reset_index()[["base_action", "n"]]
             .sort_values(["n", "base_action"], ascending=[False, True])
