+++ tests/fflogs_rotation/dawntrail/physical_ranged/test_brd.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+from ffxiv_stats.jobs import PhysicalRanged
 from pandas.testing import assert_frame_equal
 
 from crit_app.job_data.encounter_data import encounter_phases
@@ -13,43 +14,43 @@ from fflogs_rotation.job_data.data import (
 )
 from fflogs_rotation.rotation import RotationTable
 from fflogs_rotation.tests.config import headers
-from ffxiv_stats.jobs import Melee
 
 
-class TestSamuraiActions:
-    """Tests for Samurai job actions and rotations.
+class TestBardActions:
+    """Tests for Bard job actions and rotations.
 
-    Based off https://www.fflogs.com/reports/Lw7PgrFVcA1GNqnh?fight=6
+    Based of Reality's speed kill for m1s
+    https://www.fflogs.com/reports/B3gxKdn4j1W8NML9#fight=3
     """
 
     def setup_method(self):
         """Setup method to initialize common variables and RotationTable."""
-        self.report_id = "Lw7PgrFVcA1GNqnh"
-        self.fight_id = 6
+        self.report_id = "B3gxKdn4j1W8NML9"
+        self.fight_id = 3
         self.level = 100
         self.phase = 0
         self.player_id = 3
         self.pet_ids = None
 
-        self.main_stat = 5114
+        self.main_stat = 5130
         self.pet_attack_power = self.main_stat // 1.05
-        self.det = 1931
-        self.speed = 420
+        self.det = 2091
+        self.speed = 474
 
-        self.crt = 3103
-        self.dh = 2095
+        self.crt = 3177
+        self.dh = 2080
         self.wd = 146
 
-        self.delay = 2.64
+        self.delay = 3.04
         self.medication = 392
 
-        self.t = 515
+        self.t = 393.947
 
-        self.sam_rotation_7_05 = RotationTable(
+        self.brd_rotation_7_05 = RotationTable(
             headers,
             self.report_id,
             self.fight_id,
-            "Samurai",
+            "Bard",
             self.player_id,
             self.crt,
             self.dh,
@@ -67,7 +68,7 @@ class TestSamuraiActions:
             pet_ids=self.pet_ids,
         )
 
-        self.sam_analysis_7_05 = Melee(
+        self.brd_analysis_7_05 = PhysicalRanged(
             self.main_stat,
             self.det,
             self.speed,
@@ -75,53 +76,46 @@ class TestSamuraiActions:
             self.dh,
             self.wd,
             self.delay,
-            "Samurai",
             self.pet_attack_power,
             level=self.level,
         )
         # self.black_cat_ast.attach_rotation(self.black_cat_rotation.rotation_df, self.t)
 
     @pytest.fixture
-    def expected_sam_7_05_action_counts(self) -> pd.DataFrame:
+    def expected_brd_7_05_action_counts(self) -> pd.DataFrame:
         """Fixture providing expected action count data."""
         return (
             pd.DataFrame(
                 [
-                    {"base_action": "Attack", "n": 202},
-                    {"base_action": "Higanbana (tick)", "n": 159},
-                    {"base_action": "Gyofu", "n": 53},
-                    {"base_action": "Hissatsu: Shinten", "n": 51},
-                    {"base_action": "Gekko", "n": 32},
-                    {"base_action": "Kasha", "n": 27},
-                    {"base_action": "Yukikaze", "n": 27},
-                    {"base_action": "Jinpu", "n": 16},
-                    {"base_action": "Kaeshi: Setsugekka", "n": 15},
-                    {"base_action": "Midare Setsugekka", "n": 15},
-                    {"base_action": "Shoha", "n": 13},
-                    {"base_action": "Shifu", "n": 12},
-                    {"base_action": "Tendo Kaeshi Setsugekka", "n": 11},
-                    {"base_action": "Tendo Setsugekka", "n": 11},
-                    {"base_action": "Hissatsu: Senei", "n": 9},
-                    {"base_action": "Higanbana", "n": 8},
-                    {"base_action": "Hissatsu: Gyoten", "n": 7},
-                    {"base_action": "Kaeshi: Namikiri", "n": 5},
-                    {"base_action": "Ogi Namikiri", "n": 5},
-                    {"base_action": "Zanshin", "n": 5},
-                    {"base_action": "Enpi", "n": 2},
-                    {"base_action": "Hissatsu: Yaten", "n": 1},
+                    {"base_action": "Burst Shot", "n": 96},
+                    {"base_action": "Refulgent Arrow", "n": 41},
+                    {"base_action": "Heartbreak Shot", "n": 50},
+                    {"base_action": "Shot", "n": 133},
+                    {"base_action": "Pitch Perfect", "n": 19},
+                    {"base_action": "Empyreal Arrow", "n": 26},
+                    {"base_action": "Blast Arrow", "n": 6},
+                    {"base_action": "Radiant Encore", "n": 4},
+                    {"base_action": "Apex Arrow", "n": 6},
+                    {"base_action": "Resonant Arrow", "n": 4},
+                    {"base_action": "Stormbite (tick)", "n": 129},
+                    {"base_action": "Sidewinder", "n": 7},
+                    {"base_action": "Caustic Bite (tick)", "n": 128},
+                    {"base_action": "Iron Jaws", "n": 10},
+                    {"base_action": "Caustic Bite", "n": 1},
+                    {"base_action": "Stormbite", "n": 1},
                 ]
             )
             .sort_values(["n", "base_action"], ascending=[False, True])
             .reset_index(drop=True)
         )
 
-    def test_sam_7_05_action_counts(
-        self, expected_sam_7_05_action_counts: pd.DataFrame
+    def test_brd_7_05_action_counts(
+        self, expected_brd_7_05_action_counts: pd.DataFrame
     ):
         """Test that action counts match expected values for Black Cat log."""
         # Arrange
         actual_counts = (
-            self.sam_rotation_7_05.rotation_df.groupby("base_action")
+            self.brd_rotation_7_05.rotation_df.groupby("base_action")
             .sum("n")
             .reset_index()[["base_action", "n"]]
             .sort_values(["n", "base_action"], ascending=[False, True])
@@ -129,4 +123,4 @@ class TestSamuraiActions:
         )
 
         # Assert
-        assert_frame_equal(actual_counts, expected_sam_7_05_action_counts)
+        assert_frame_equal(actual_counts, expected_brd_7_05_action_counts)
