+++ tests/fflogs_rotation/dawntrail/tank/test_war.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+from ffxiv_stats.jobs import Tank
 from pandas.testing import assert_frame_equal
 
 from crit_app.job_data.encounter_data import encounter_phases
@@ -13,44 +14,44 @@ from fflogs_rotation.job_data.data import (
 )
 from fflogs_rotation.rotation import RotationTable
 from fflogs_rotation.tests.config import headers
-from ffxiv_stats.jobs import MagicalRanged
 
 
-class TestBlackMageActions:
-    """Tests for Black Mage job actions and rotations.
+class TestWarriorActions:
+    """Tests for Warrior job actions and rotations.
 
-    Based off
-    https://www.fflogs.com/reports/yWvADHcK8XJCj4pF?fight=19
+    Based off https://www.fflogs.com/reports/2PhAr7tRc8GpKDNa?fight=2
     """
 
     def setup_method(self):
         """Setup method to initialize common variables and RotationTable."""
-        self.report_id = "yWvADHcK8XJCj4pF"
-        self.fight_id = 19
+        self.report_id = "2PhAr7tRc8GpKDNa"
+        self.fight_id = 2
         self.level = 100
         self.phase = 0
-        self.player_id = 104
+        self.player_id = 5
         self.pet_ids = None
 
-        self.main_stat = 5126
+        self.main_stat = 5079
         self.pet_attack_power = self.main_stat // 1.05
-        self.det = 1572
-        self.speed = 1047
+        self.det = 2525
+        self.speed = 420
 
-        self.crt = 3321
-        self.dh = 1882
+        self.crt = 3253
+        self.dh = 1176
         self.wd = 146
 
-        self.delay = 3.28
+        self.ten = 868
+
+        self.delay = 3.36
         self.medication = 392
 
-        self.t = 432
+        self.t = 545
 
-        self.blm_rotation_7_05 = RotationTable(
+        self.war_rotation_7_05 = RotationTable(
             headers,
             self.report_id,
             self.fight_id,
-            "BlackMage",
+            "Warrior",
             self.player_id,
             self.crt,
             self.dh,
@@ -68,49 +69,54 @@ class TestBlackMageActions:
             pet_ids=self.pet_ids,
         )
 
-        self.blm_analysis_7_05 = MagicalRanged(
+        self.war_analysis_7_05 = Tank(
             self.main_stat,
-            400,
             self.det,
             self.speed,
+            self.ten,
             self.crt,
             self.dh,
             self.wd,
             self.delay,
+            "Warrior",
             self.pet_attack_power,
             level=self.level,
         )
         # self.black_cat_ast.attach_rotation(self.black_cat_rotation.rotation_df, self.t)
 
     @pytest.fixture
-    def expected_blm_7_05_action_counts(self) -> pd.DataFrame:
+    def expected_war_7_05_action_counts(self) -> pd.DataFrame:
         """Fixture providing expected action count data."""
         return (
             pd.DataFrame(
                 [
-                    {"base_action": "Fire IV", "n": 73},
-                    {"base_action": "Paradox", "n": 20},
-                    {"base_action": "Xenoglossy", "n": 18},
-                    {"base_action": "High Thunder", "n": 15},
-                    {"base_action": "Despair", "n": 13},
-                    {"base_action": "Flare Star", "n": 11},
-                    {"base_action": "Fire III", "n": 10},
-                    {"base_action": "Blizzard IV", "n": 8},
-                    {"base_action": "Blizzard III", "n": 8},
-                    {"base_action": "High Thunder (tick)", "n": 143},
+                    {"base_action": "Attack", "n": 159},
+                    {"base_action": "Fell Cleave", "n": 48},
+                    {"base_action": "Heavy Swing", "n": 44},
+                    {"base_action": "Maim", "n": 43},
+                    {"base_action": "Storm's Path", "n": 27},
+                    {"base_action": "Onslaught", "n": 20},
+                    {"base_action": "Upheaval", "n": 18},
+                    {"base_action": "Inner Chaos", "n": 16},
+                    {"base_action": "Storm's Eye", "n": 16},
+                    {"base_action": "Primal Ruination", "n": 9},
+                    {"base_action": "Primal Rend", "n": 9},
+                    {"base_action": "Primal Wrath", "n": 9},
+                    {"base_action": "Damnation", "n": 32},
+                    {"base_action": "Tomahawk", "n": 2},
                 ]
             )
             .sort_values(["n", "base_action"], ascending=[False, True])
             .reset_index(drop=True)
         )
 
-    def test_blm_7_05_action_counts(
-        self, expected_blm_7_05_action_counts: pd.DataFrame
+    def test_war_7_05_action_counts(
+        self, expected_war_7_05_action_counts: pd.DataFrame
     ):
-        """Test that action counts match expected values for 7.05 log."""
+        """Test that action counts match expected values against FFlogs."""
         # Arrange
         actual_counts = (
-            self.blm_rotation_7_05.rotation_df.groupby("base_action")
+            self.war_rotation_7_05.rotation_df.groupby("base_action")
             .sum("n")
             .reset_index()[["base_action", "n"]]
             .sort_values(["n", "base_action"], ascending=[False, True])
@@ -118,4 +124,4 @@ class TestBlackMageActions:
         )
 
         # Assert
-        assert_frame_equal(actual_counts, expected_blm_7_05_action_counts)
+        assert_frame_equal(actual_counts, expected_war_7_05_action_counts)
