+++ tests/fflogs_rotation/test_black_mage.py
