+++ tests/fflogs_rotation/test_black_mage_7_2.py
