+++ tests/fflogs_rotation/test_monk.py
@@ -1,4 +1,4 @@
-from fflogs_rotation.monk import MonkActions
+# from fflogs_rotation.monk import MonkActions
 
 # TODO: set up monk test
