+++ tests/fflogs_rotation/test_ninja.py
